Ensure sessions cresated by JSchSshClient are closed
This patch changes behaviour of JSchSshClient so sessions are created on
per process/upload basis and closed immediatyly after procss/upload is
completed.

Additionally we set JSch threads as daemon so JVM would still exit if
some JSch threads are hanging around. This can be an issue in example
when test timeout occurs.

Reviewers: gkokosinski